requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.3
pydantic-settings==2.3.0
google-generativeai==0.6.0
python-json-logger==2.0.7
python-jose==3.3.0
tenacity==8.2.3
pytest==8.2.1
pytest-asyncio==0.23.6
beautifulsoup4==4.12.3
chromadb==0.5.0
aiohttp==3.9.5
google-search-results==2.4.2
scipy==1.13.0
numpy==1.26.4